feat: update widget and add to page
