+++ src/ros/nova-gui/nova-gui/src/ros/services/rosServiceMessages.ts
@@ -33,8 +33,6 @@ export const rosServiceMessages = {
   [RosService.THETA_360_CAM_CAPTURE]: "std_srvs/srv/Trigger",
   [RosService.CALL_RAMAN_SPEC]: "science_interfaces/srv/RamanSpec",
   [RosService.CALL_RAMAN_MECH]: "science_interfaces/srv/RamanMech",
-  [RosService.UV_VIS_LED_1]: "std_srvs/srv/SetBool",
-  [RosService.UV_VIS_LED_2]: "std_srvs/srv/SetBool",
   [RosService.LEDS]: "science_interfaces/srv/Toggle",
   [RosService.CACHE_1]: "science_interfaces/srv/CacheCommand",
   [RosService.CACHE_2]: "science_interfaces/srv/CacheCommand",
