+++ src/ros/nova-gui/nova-gui/src/ros/services/rosServiceTypes.ts
@@ -153,14 +153,6 @@ export interface RosServiceInterface {
     IRosScienceInterfacesRamanMechRequest,
     IRosScienceInterfacesRamanMechResponse
   >;
-  [RosService.UV_VIS_LED_1]: RosServiceMessage<
-    IRosStdSrvsSetBoolRequest,
-    IRosStdSrvsSetBoolResponse
-  >;
-  [RosService.UV_VIS_LED_2]: RosServiceMessage<
-    IRosStdSrvsSetBoolRequest,
-    IRosStdSrvsSetBoolResponse
-  >;
   [RosService.LEDS]: RosServiceMessage<
     IRosScienceInterfacesToggleRequest,
     IRosScienceInterfacesToggleResponse
